Fix hook bug in sheets refunds (#1657)
